synchronising with xbmc-pvr-addons repository

+++ src/lib/tsreader/MultiFileReader.cpp
@@ -41,7 +41,6 @@
 
 #if !defined(TARGET_WINDOWS)
 #include <sys/time.h>
-#include "PlatformDefs.h"
 #include "os-dependent.h"
 #define SUCCEEDED(hr) (((HRESULT)(hr)) >= 0)
 #endif
@@ -353,7 +352,7 @@ long MultiFileReader::RefreshTSBufferFile()
   
     pBuffer = (wchar_t*) new char[(unsigned int)remainingLength];
 
-    result=m_TSBufferFile.Read((unsigned char*)pBuffer, (ULONG)remainingLength, &bytesRead);
+    result = m_TSBufferFile.Read((unsigned char*) pBuffer, (uint32_t)remainingLength, &bytesRead);
     if (!SUCCEEDED(result)||  (int64_t)bytesRead != remainingLength) Error=0x20;
 
     //unsigned char* pb = (unsigned char*) pBuffer;
@@ -372,10 +371,10 @@ long MultiFileReader::RefreshTSBufferFile()
       Error |= 0x40;
 
     if(Error == 0)
-	  {
-		  filesAdded2 = *((TSBUFFERLONG*)(readBuffer + 0));
-		  filesRemoved2 = *((TSBUFFERLONG*)(readBuffer + sizeof(filesAdded2)));
-	  }
+    {
+      filesAdded2 = *((TSBUFFERLONG*)(readBuffer + 0));
+      filesRemoved2 = *((TSBUFFERLONG*)(readBuffer + sizeof(filesAdded2)));
+    }
 
     delete[] readBuffer;
 
@@ -392,10 +391,11 @@ long MultiFileReader::RefreshTSBufferFile()
       usleep(5000);
     }
 
-    if (Error) delete[] pBuffer;
+    if (Error)
+      delete[] pBuffer;
 
-    Loop-- ;
-  } while ( Error && Loop ) ; // If Error is set, try again...until Loop reaches 0.
+    Loop--;
+  } while ( Error && Loop ); // If Error is set, try again...until Loop reaches 0.
  
   if (Loop < 8)
   {
@@ -516,7 +516,7 @@ long MultiFileReader::RefreshTSBufferFile()
       // Move the wchar buffer pointer to the next wchar string
 #if defined(TARGET_WINDOWS)
       pwCurrFile += (length + 1);
-#elif defined(TARGET_LINUX) || defined(TARGET_OSX)
+#elif defined(TARGET_LINUX) || defined(TARGET_DARWIN)
       unsigned short *pus = (unsigned short *) pwCurrFile;
       pus += (length + 1);
       pwCurrFile = (wchar_t *) pus;
@@ -651,7 +651,7 @@ long MultiFileReader::GetFileLength(const char* pFilename, int64_t &length)
     return HRESULT_FROM_WIN32(dwErr);
   }
   return S_OK;
-#elif defined(TARGET_LINUX) || defined(TARGET_OSX)
+#elif defined(TARGET_LINUX) || defined(TARGET_DARWIN)
   //USES_CONVERSION;
 
   length = 0;
@@ -691,7 +691,7 @@ size_t MultiFileReader::WcsLen(const void *str)
 {
 #if defined(TARGET_WINDOWS)
   return wcslen((const wchar_t *)str);
-#elif defined(TARGET_LINUX) || defined(TARGET_OSX)
+#elif defined(TARGET_LINUX) || defined(TARGET_DARWIN)
   const unsigned short *eos = (const unsigned short*)str;
   while( *eos++ ) ;
   return( (size_t)(eos - (const unsigned short*)str) -1);
@@ -705,7 +705,7 @@ size_t MultiFileReader::WcsToMbs(char *s, const void *w, size_t n)
 {
 #if defined(TARGET_WINDOWS)
   return wcstombs(s, (const wchar_t *)w, n);
-#elif defined(TARGET_LINUX) || defined(TARGET_OSX)
+#elif defined(TARGET_LINUX) || defined(TARGET_DARWIN)
   size_t i = 0;
   const unsigned short *wc = (const unsigned short*) w;
   while(wc[i] && (i < n))
